# swagger.yaml
swagger: '2.0'
info:
  description: '"Swagger configuration for application "'
  version: 1.1.0
  title: Bookstore - Spring Boot Swagger Configuration
#  contact:
#    name: Bookstore
#    url: https://www.youtube.com/channel/UCMpJ8m1w9t7EFcF9x8rs02A
#    email: baljibeka@gmail.com
#  license:
#    name: Apache 2.0
#    url: https://www.apache.org/licenses/LICENSE-2.0"
host: localhost:8080
basePath: /
...
paths:
...
  /publishers:
    get:
      tags:
        - publisher-controller
      summary: getAll
      operationId: getAllUsingGET_4
      produces:
        - '*/*'
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/PublisherShortDTO'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /publishers/publisher:
    post:
      tags:
        - publisher-controller
      summary: savePublisher
      operationId: savePublisherUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: publisherShortDTO
          description: publisherShortDTO
          required: true
          schema:
            $ref: '#/definitions/PublisherShortDTO'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/PublisherShortDTO'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    put:
      tags:
        - publisher-controller
      summary: updatePublisher
      operationId: updatePublisherUsingPUT
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: publisherDTO
          description: publisherDTO
          required: true
          schema:
            $ref: '#/definitions/PublisherDTO'
      responses:
        '200':
          description: OK
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /publishers/publisher/name/{publisherName}:
    get:
      tags:
        - publisher-controller
      summary: getPublisherByName
      operationId: getPublisherByNameUsingGET
      produces:
        - '*/*'
      parameters:
        - name: publisherName
          in: path
          description: publisherName
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/PublisherShortDTO'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /publishers/publisher/{publisherID}:
    get:
      tags:
        - publisher-controller
      summary: getPublisherById
      operationId: getPublisherByIdUsingGET
      produces:
        - '*/*'
      parameters:
        - name: publisherID
          in: path
          description: publisherID
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/PublisherShortDTO'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    delete:
      tags:
        - publisher-controller
      summary: deletePublisherById
      operationId: deletePublisherByIdUsingDELETE
      produces:
        - '*/*'
      parameters:
        - name: publisherID
          in: path
          description: publisherID
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  /users:
    get:
      tags:
        - user-controller
      summary: getAll
      operationId: getAllUsingGET_5
      produces:
        - '*/*'
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/UserDTO'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /users/user:
    post:
      tags:
        - user-controller
      summary: create
      operationId: createUsingPOST_2
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: userFullDTO
          description: userFullDTO
          required: true
          schema:
            $ref: '#/definitions/UserFullDTO'
      responses:
        '200':
          description: OK
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    put:
      tags:
        - user-controller
      summary: update
      operationId: updateUsingPUT_2
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: userFullDTO
          description: userFullDTO
          required: true
          schema:
            $ref: '#/definitions/UserFullDTO'
      responses:
        '200':
          description: OK
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /users/user/admin:
    put:
      tags:
        - user-controller
      summary: adminUpdate
      operationId: adminUpdateUsingPUT
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: userAdminDTO
          description: userAdminDTO
          required: true
          schema:
            $ref: '#/definitions/UserAdminDTO'
      responses:
        '200':
          description: OK
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /users/user/{userID}:
    get:
      tags:
        - user-controller
      summary: findByID
      operationId: findByIDUsingGET
      produces:
        - '*/*'
      parameters:
        - name: userID
          in: path
          description: userID
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/UserDTO'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    delete:
      tags:
        - user-controller
      summary: delete
      operationId: deleteUsingDELETE_2
      produces:
        - '*/*'
      parameters:
        - name: userID
          in: path
          description: userID
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
definitions:
...
  PublisherDTO:
    type: object
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
      publishedBooks:
        type: array
        items:
          type: integer
          format: int64
    title: PublisherDTO
  PublisherShortDTO:
    type: object
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
    title: PublisherShortDTO
  UserAdminDTO:
    type: object
    properties:
      id:
        type: integer
        format: int64
      isBlocked:
        type: boolean
      role:
        type: string
        enum:
          - ADMIN
          - USER
    title: UserAdminDTO
  UserDTO:
    type: object
    properties:
      id:
        type: integer
        format: int64
      login:
        type: string
      role:
        type: string
        enum:
          - ADMIN
          - USER
    title: UserDTO
  UserFullDTO:
    type: object
    properties:
      id:
        type: integer
        format: int64
      login:
        type: string
      password:
        type: string
    title: UserFullDTO